Move coverage tests to unit tests in conversion.rs
