--- tests/coverage_tests.rs
@@ -1,569 +1,0 @@
-//! Additional tests to improve code coverage for semantic conversion
-//!
-//! These tests target specific edge cases and error conditions in
-//! `src/semantic/conversion.rs` that might be missed by the main test suite.
-
-use glossa::ast::build_ast;
-use glossa::semantic::analyze_program;
-
-/// Helper to compile source and check for specific error messages
-fn compile_and_expect_error(source: &str, error_fragment: &str) {
-    let ast = build_ast(source).expect("AST build failed");
-    match analyze_program(&ast) {
-        Ok(_) => panic!(
-            "Expected error containing '{}', but analysis succeeded",
-            error_fragment
-        ),
-        Err(e) => {
-            let error_msg = format!("{}", e);
-            assert!(
-                error_msg.contains(error_fragment),
-                "Expected error '{}' to contain '{}'",
-                error_msg,
-                error_fragment
-            );
-        }
-    }
-}
-
-/// Helper to compile source successfully
-fn compile_success(source: &str) {
-    let ast = build_ast(source).expect("AST build failed");
-    if let Err(e) = analyze_program(&ast) {
-        panic!("Analysis failed for '{}': {}", source, e);
-    }
-}
-
-#[test]
-fn test_assignment_logic() {
-    // Valid assignment
-    // ξ = 5; ξ = 10;
-    // Note: Variable must be mutable (μετά) to be assigned
-    let source = "ξ πέντε μετά ἔστω. ξ δέκα γίγνεται.";
-    compile_success(source);
-}
-
-#[test]
-fn test_assignment_error_undefined() {
-    // Assigning to undefined variable
-    let source = "ξ δέκα γίγνεται.";
-    // "The 'xi' was not defined - first define it"
-    // Greek error message: "Τὸ «ξ» οὐχ ὡρίσθη"
-    compile_and_expect_error(source, "οὐχ ὡρίσθη");
-}
-
-#[test]
-fn test_assignment_error_immutable() {
-    // Assigning to immutable variable
-    let source = "ξ πέντε ἔστω. ξ δέκα γίγνεται.";
-    // "immutable variable"
-    // Greek error message likely involves "ἀμετάβλητον" or similar from `immutable_assignment`
-    // Let's match on the variable name which should be in the error
-    compile_and_expect_error(source, "ξ");
-}
-
-#[test]
-fn test_assignment_error_no_value() {
-    // Assignment without a value (e.g., just "x becomes")
-    // This is hard to construct syntactically as the parser might reject it first,
-    // but if we have just subject and verb...
-    // "ξ γίγνεται."
-    let source = "ξ πέντε μετά ἔστω. ξ γίγνεται.";
-    // "By the action 'xi becomes' a value is needed"
-    // Greek: "Τῇ πράξει «ξ γίγνεται» δεῖ τιμῆς"
-    compile_and_expect_error(source, "δεῖ τιμῆς");
-}
-
-#[test]
-fn test_binding_swapped_subject_object() {
-    // "5 let x" instead of "let x 5"
-    // This triggers the heuristic in detect_variable_binding
-    // Standard: ξ πέντε ἔστω.
-    // Swapped: πέντε ξ ἔστω. (where xi is the variable)
-    // Note: literals can't be subjects usually, so we need two identifiers?
-    // Or maybe just ensure normal binding works
-    let source = "ξ πέντε ἔστω.";
-    compile_success(source);
-}
-
-#[test]
-fn test_collection_instantiation() {
-    // HashSet instantiation
-    // "s new Set let."
-    let source = "σ νέον σύνολον ἔστω.";
-    compile_success(source);
-
-    // HashMap instantiation
-    // "m new Map let."
-    let source = "χ νέον χάρτης ἔστω.";
-    compile_success(source);
-}
-
-#[test]
-fn test_binding_with_false_participle() {
-    // Variable name that looks like a participle but isn't
-    // "topikon" (local) ends in -on which is a participle ending
-    // but here it is a variable name
-    let source = "τοπικον πέντε ἔστω. τοπικον λέγε.";
-    compile_success(source);
-}
-
-// Additional tests for missing coverage
-
-#[test]
-fn test_pop_push_insert() {
-    // pop
-    let source_pop = "ξ [1, 2] ἔστω. ξ ἕλκεται.";
-    compile_success(source_pop);
-
-    // push
-    let source_push = "ξ [] ἔστω. ξ ὠθεῖ 1.";
-    compile_success(source_push);
-
-    // insert (set)
-    let source_insert_set = "σ νέον σύνολον ἔστω. σ 1 τίθησι.";
-    compile_success(source_insert_set);
-
-    // insert (map)
-    // Corrected to use two values for map insert
-    let source_insert_map = "χ νέον χάρτης ἔστω. χ «κλειδί» 2 τίθησι.";
-    compile_success(source_insert_map);
-}
-
-#[test]
-fn test_comparison_subjunctive() {
-    // Comparison with subjunctive verb form (implies "if")
-    // "x 5 greater be?" (subjunctive)
-    // "εἰ ξ πέντε μεῖζον ᾖ" - "if xi is greater than 5"
-    // We need a context where this is an expression, e.g., in a print or if
-    let source = "ξ πέντε ἔστω. εἰ ξ πέντε μεῖζον ᾖ, «μείζον» λέγε.";
-    compile_success(source);
-}
-
-#[test]
-fn test_custom_struct_instantiation() {
-    // Define a struct and instantiate it
-    // Syntax: εἶδος Name ὁρίζειν { field type. }.
-    let source = "
-    εἶδος Σημεῖον ὁρίζειν { χ ἀριθμοῦ. }.
-    π νέον Σημεῖον 1 ἔστω.
-    ";
-    compile_success(source);
-}
-
-#[test]
-fn test_print_variants() {
-    // Print with operator
-    // Using simple literals that shouldn't be confused with verbs
-    let source_op = "1 καὶ 2 λέγε.";
-    compile_success(source_op);
-
-    // Print with unwrap (!)
-    // "xi something 5 let" -> let xi = Some(5)
-    let source_unwrap = "ξ τί 5 ἔστω. ξ! λέγε.";
-    compile_success(source_unwrap);
-}
-
-#[test]
-fn test_function_call_patterns() {
-    // Call user defined function
-    // Syntax: ἔργον name(args) δὸς result.
-    let source = "
-    ἔργον φ(χ) δὸς χ.
-    ξ φ(5) ἔστω.
-    ";
-    compile_success(source);
-}
-
-#[test]
-fn test_binding_propagation() {
-    // Binding with propagation operator (;)
-    // "xi phi(5) let?" -> let xi = phi(5)?;
-    let source = "
-    ἔργον φ(χ) δὸς ἐπιτυχία χ.
-    ξ φ(5) ἔστω;
-    ";
-    compile_success(source);
-}
-
-#[test]
-fn test_print_string_split() {
-    // String split method call
-    let source = "ξ «α-β» ἔστω. ξ κατὰ «-» σχίζεται λέγε.";
-    compile_success(source);
-}
-
-#[test]
-fn test_comparison_unknown_var() {
-    // Comparison with unknown variable (fallback to false logic)
-    // "If unknown 5 greater be" -> should not panic, though semantics are weird
-    // The code generates BooleanLiteral(false) for unknown vars in this specific check
-    let source = "εἰ αγνωστον πέντε μεῖζον ᾖ, «ναι» λέγε.";
-    compile_success(source);
-}
-
-#[test]
-fn test_binding_error_no_subject() {
-    // Binding verb without subject
-    // "let."
-    let source = "ἔστω.";
-    compile_and_expect_error(source, "Binding without subject");
-}
-
-#[test]
-fn test_struct_instantiation_fallthrough() {
-    // Structure that looks like struct instantiation but adjective isn't "new"
-    // "xi positive five let." -> xi = 5 (ignoring positive? or treating as value?)
-    // Actually, detect_struct_instantiation checks for "νεος".
-    // If it fails, it falls through to binding.
-    // Binding takes "positive five" as value.
-    let source = "ξ θετικά πέντε ἔστω.";
-    compile_success(source);
-}
-
-#[test]
-fn test_print_with_operator_and_subject() {
-    // Print "xi + 5"
-    let source = "ξ πέντε ἔστω. ξ καὶ 5 λέγε.";
-    compile_success(source);
-}
-
-#[test]
-fn test_binding_default_zero() {
-    // Binding with no value provided should default to 0
-    // "xi let."
-    let source = "ξ ἔστω.";
-    compile_success(source);
-}
-
-#[test]
-fn test_binding_result_object() {
-    // Binding with Result in object slot
-    // "xi success 5 let" -> let xi = Ok(5)
-    // "επιτυχια" is the object, "5" is literal
-    let source = "ξ ἐπιτυχία 5 ἔστω.";
-    compile_success(source);
-}
-
-#[test]
-fn test_query_literals_only() {
-    // Query with no subject
-    // "5?" (Is 5?) - just evaluates expression and prints it (conceptually)
-    let source = "5?";
-    compile_success(source);
-}
-
-#[test]
-fn test_function_call_in_object_slot() {
-    // Attempt to put function call in object slot
-    // "xi phi 5 let" where phi is a function
-    // The assembler might treat phi as subject or object depending on case
-    // We assume default/nominative is OK for function names
-    let source = "
-    ἔργον φ(χ) δὸς χ.
-    ξ φ 5 ἔστω.
-    ";
-    compile_success(source);
-}
-
-#[test]
-fn test_binding_unwrap() {
-    // Bind unwrap
-    let source = "ξ τί 5 ἔστω. ψ ξ! ἔστω.";
-    compile_success(source);
-}
-
-#[test]
-fn test_binding_array() {
-    // Bind array
-    let source = "ξ [1, 2] ἔστω.";
-    compile_success(source);
-}
-
-#[test]
-fn test_binding_property() {
-    // Bind property access
-    // "struct S {x num.} let s = new S 5. let y = s.x."
-    let source = "
-    εἶδος Σημεῖον ὁρίζειν { χ ἀριθμοῦ. }.
-    π νέον Σημεῖον 1 ἔστω.
-    ξ που χ ἔστω.
-    ";
-    compile_success(source);
-}
-
-#[test]
-fn test_binding_result_nominative() {
-    // Binding with Result in nominative slot (extra nominative)
-    // "xi 5 success let" -> xi=sub, 5=lit, success=nom
-    let source = "ξ 5 ἐπιτυχία ἔστω.";
-    compile_success(source);
-}
-
-#[test]
-fn test_function_call_genitive() {
-    // Function call with name in genitive
-    // "xi phi-of 5 let" -> xi=sub, phi=gen, 5=lit
-    let source = "
-    ἔργον λόγος(χ) δὸς χ.
-    ξ λόγου 5 ἔστω.
-    ";
-    compile_success(source);
-}
-
-#[test]
-fn test_binding_subject_special_words() {
-    // "ouden 5 let" -> let ouden = None (ignoring 5?)
-    // This hits `extract_value` subject check for None/Some
-    // "ouden" (nothing) as a variable name is allowed by parser?
-    let source_none = "ουδεν 5 ἔστω.";
-    compile_success(source_none);
-
-    // "ti 5 let" -> let ti = Some(5)
-    let source_some = "τι 5 ἔστω.";
-    compile_success(source_some);
-}
-
-#[test]
-fn test_binary_op_in_argument() {
-    // Function call with binary expression arg
-    // "phi(1 + 2) let"
-    // This calls `classify_value_expression` recursively
-    let source = "
-    ἔργον φ(χ) δὸς χ.
-    ξ φ(1 καὶ 2) ἔστω.
-    ";
-    compile_success(source);
-}
-
-#[test]
-fn test_property_access_on_non_struct() {
-    // "xi 5 let. of-xi 5 say." -> owner is xi (number), not struct.
-    // Should NOT match property access logic.
-    // Should fall through to print default.
-    let source = "ξ 5 ἔστω. που ξ λέγε.";
-    compile_success(source);
-}
-
-#[test]
-fn test_struct_instantiation_unknown_type() {
-    // "xi new Unknown 5 let."
-    // Unknown type should fall through to variable binding.
-    // Variable "xi" bound to "new Unknown 5" (or whatever parts remain).
-    // Assembler might parse "Unknown" as a variable name or something.
-    // Important is that `detect_struct_instantiation` returns `Ok(None)`.
-    let source = "ξ νέον ἄγνωστον 5 ἔστω.";
-    compile_success(source);
-}
-
-#[test]
-fn test_print_subject_and_object() {
-    // "xi 5 let. psi 10 let. xi psi say."
-    // Print with both subject and object variables.
-    let source = "ξ 5 ἔστω. ψ 10 ἔστω. ξ ψ λέγε.";
-    compile_success(source);
-}
-
-#[test]
-fn test_expression_propagation() {
-    // Expression statement with propagation (semicolon)
-    // "phi(5);" -> `phi(5)?`
-    // Requires function returning Result.
-    let source = "
-    ἔργον φ(χ) δὸς ἐπιτυχία χ.
-    φ(5);
-    ";
-    compile_success(source);
-}
-
-#[test]
-fn test_extract_value_result_err() {
-    // "xi error 5 let." -> let xi = Err(5)
-    let source = "ξ σφάλμα 5 ἔστω.";
-    compile_success(source);
-}
-
-#[test]
-fn test_function_call_nominative_extra() {
-    // "xi yi phi 5 let." -> xi=sub, yi=obj, phi=nom, 5=lit
-    // Function call where function name is in nominatives list
-    // Let's define psi first to be safe.
-    let source_safe = "
-    ἔργον φ(χ) δὸς χ.
-    ψ 10 ἔστω.
-    ξ ψ φ 5 ἔστω.
-    ";
-    compile_success(source_safe);
-}
-
-#[test]
-fn test_extract_value_literals_ops() {
-    // "xi 1 + 2 let." -> let xi = 1 + 2
-    let source = "ξ 1 καὶ 2 ἔστω.";
-    compile_success(source);
-}
-
-#[test]
-fn test_extract_value_object_literal_op() {
-    // "xi psi + 2 let." -> let xi = psi + 2
-    let source = "ψ 10 ἔστω. ξ ψ καὶ 2 ἔστω.";
-    compile_success(source);
-}
-
-#[test]
-fn test_iterator_fallthrough() {
-    // "xi unknown-participle say."
-    // Should fall through `detect_iterator_pattern_statement`
-    // and go to `detect_print`.
-    // "unknown-participle" needs to be something that looks like a participle?
-    // Or just a word that the assembler tagged as participle?
-    // Assembler only tags known participles?
-    // Use a known participle that is NOT a map/filter/find verb?
-    // Is there one? Most participles are tied to operations.
-    // Maybe we can rely on `detect_iterator_pattern` returning None if the verb is missing?
-    // "xi doubled say" (no verb?) -> "say" is verb.
-    // If "doubled" is participle.
-    // `detect_iterator_pattern` checks if "doubled" maps to a known op.
-    // If we use a "false" participle (like variable name ending in -on).
-    // "xi topikon say." -> "topikon" looks like participle.
-    // `detect_iterator_pattern` might see it.
-    // If it returns None, we fall through.
-    let source = "ξ 5 ἔστω. τοπικον λέγε.";
-    compile_success(source);
-}
-
-#[test]
-fn test_property_access_print_struct() {
-    // Test property access print on a struct
-    // Pattern: genitive_var nominative_field say.
-    // "struct S {x num.} let s = new S 5. of-s x say."
-    let source = "
-    εἶδος Σημεῖον ὁρίζειν { χ ἀριθμοῦ. }.
-    σ νέον Σημεῖον 5 ἔστω.
-    σημείου χ λέγε.
-    ";
-    compile_success(source);
-}
-
-#[test]
-fn test_classify_value_expr_error() {
-    // Attempt: "Binary operation missing right operand"
-    // "phi(1 + ) let"
-    // Parser likely catches this, but if we can create a state where it passes:
-    // "phi(1 +) let"
-    // If we use a trailing operator in a place where an expression is expected?
-    // Hard to construct. Skipped for now.
-}
-
-#[test]
-fn test_struct_instantiation_collections() {
-    // Explicitly test HashSet and HashMap instantiation
-    // HashSet
-    let source_set = "σ νέον σύνολον ἔστω.";
-    compile_success(source_set);
-
-    // HashMap
-    let source_map = "χ νέον χάρτης ἔστω.";
-    compile_success(source_map);
-}
-
-#[test]
-fn test_struct_instantiation_failure() {
-    // Valid binding verb, "new" adjective, but unknown type
-    // Should fall through to variable binding
-    // "xi new UnknownType 5 let"
-    let source = "ξ νέον ΆγνωστοΤύπο 5 ἔστω.";
-    compile_success(source);
-}
-
-#[test]
-fn test_binding_from_index_access() {
-    // "xi arr[0] let"
-    let source = "
-    α [1, 2] ἔστω.
-    ξ α[0] ἔστω.
-    ";
-    compile_success(source);
-}
-
-#[test]
-fn test_string_methods_join_len() {
-    // "s 'a,b' let. s split ',' join '-' say."
-    // Split returns iterator (list), join works on list.
-    // Wait, split returns List<String>. Join is a method on List<String>?
-    // Or is join a method on String that takes a list?
-    // In Rust: `s.split(",").collect::<Vec<_>>().join("-")`
-    // Glossa mapping:
-    // split -> `split` method on String.
-    // join -> `join` method on List<String>.
-    // "s split ',' join '-' say."
-    let source = "
-    σ «α,β» ἔστω.
-    σ κατὰ «,» σχίζεται μὲ «-» ἑνώνεται λέγε.
-    ";
-    compile_success(source);
-
-    // len
-    // "s len say."
-    let source_len = "
-    σ «αβγ» ἔστω.
-    σ μῆκος λέγε.
-    ";
-    compile_success(source_len);
-}
-
-#[test]
-fn test_nested_function_call_phrases() {
-    // "xi phi((psi(5))) let."
-    // Nested parenthesis
-    let source = "
-    ἔργον ψ(χ) δὸς χ.
-    ἔργον φ(χ) δὸς χ.
-    ξ φ( (ψ(5)) ) ἔστω.
-    ";
-    compile_success(source);
-}
-
-#[test]
-fn test_binding_empty_array() {
-    // "xi [] let."
-    let source = "ξ [] ἔστω.";
-    compile_success(source);
-}
-
-#[test]
-fn test_binding_swapped_identifiers() {
-    // Test the heuristic for swapped subject/object
-    // "xi yi let" where one is defined and one is not
-    // If 'xi' is defined and 'yi' is not, it might assume 'yi' is the variable
-    // "xi 5 let. xi yi let." -> let yi = xi
-    let source = "
-    ξ 5 ἔστω.
-    ξ ψ ἔστω.
-    ";
-    compile_success(source);
-}
-
-#[test]
-fn test_array_literals_mixed() {
-    // "xi [true, 'str'] let."
-    // Hits classify_value_expression literal branches
-    // Note: Glossa might not support mixed arrays easily, but `classify_value_expression` handles individual literals.
-    // Let's test them separately to be safe.
-    let source_bool = "ξ [«ναι»] ἔστω.";
-    compile_success(source_bool);
-
-    let source_str = "ξ [«κείμενον»] ἔστω.";
-    compile_success(source_str);
-}
-
-#[test]
-fn test_struct_instantiation_multi_field() {
-    // Struct with multiple fields
-    // "struct P {x num. y num.} let p = new P 1 2."
-    let source = "
-    εἶδος Σημεῖον ὁρίζειν { χ ἀριθμοῦ. ψ ἀριθμοῦ. }.
-    π νέον Σημεῖον 1 2 ἔστω.
-    ";
-    compile_success(source);
-}
